requirements.txt
git+https://github.com/Rapptz/asqlite.git
discord.py
python-dotenv
jishaku
